Fix hgss sound slots species reading
2 bytes is enough for species ID
for hgss the species ID are stored every 2 bytes instead of 4
bytes(dppt)

+++ PKHeX/Legality/Structures/EncounterArea.cs
@@ -380,7 +380,7 @@ namespace PKHeX.Core
                 if (baseSlot.LevelMin <= 0)
                     continue;
 
-                int species = BitConverter.ToInt32(data, ofs + i * slotSize);
+                int species = BitConverter.ToUInt16(data, ofs + i * slotSize);
                 if (species <= 0)
                     continue;
